# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: EasyEscrow.ai API
  version: 1.0.0
  description: Non-custodial atomic swap platform for NFTs and SOL on Solana, plus institutional cross-border USDC escrow payments. Browse the Documentation section for guides, or jump to API Reference for endpoint details.
  contact:
    name: EasyEscrow.ai Support
    email: support@easyescrow.ai
  license:
    name: ISC
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://staging-api.easyescrow.ai
    description: Staging environment (Devnet)
  - url: https://api.easyescrow.ai
    description: Production (Mainnet)
...
paths:
...
  /api/v1/institution/settings:
    get:
      tags:
        - Institution Settings
      summary: Get client settings
      description: Retrieve the authenticated client's settings including wallets and configuration.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Settings retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/InstitutionSettings'
                  timestamp:
                    type: string
                    format: date-time
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - Institution Settings
      summary: Update client settings
      description: Update general settings for the authenticated client (e.g., company name, webhook URL).
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InstitutionSettingsUpdate'
            example:
              companyName: "Acme Corporation Pty Ltd"
              webhookUrl: "https://acmecorp.com/webhooks/escrow"
      responses:
        '200':
          description: Settings updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/InstitutionSettings'
                  timestamp:
                    type: string
                    format: date-time
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/institution/settings/wallets:
    put:
      tags:
        - Institution Settings
      summary: Update wallet addresses
      description: Update the Solana wallet addresses associated with the authenticated client.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InstitutionWalletsUpdate'
            example:
              payerWallet: "498GViCLvzbGnRoByJCAj7skXkAe3NBpCY2Wghcd2e4R"
              recipientWallet: "8LL197pziojWHtS3zeyJonrh1swKvMZpumfesVmDgUcZ"
      responses:
        '200':
          description: Wallets updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    description: Updated wallet configuration
                  timestamp:
                    type: string
                    format: date-time
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/institution/api-keys:
    post:
      tags:
        - Institution Settings
      summary: Generate API key
      description: |
        Generate a new API key for programmatic access. The full key is returned **only once** in
        the response -- store it securely.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InstitutionApiKeyCreate'
            example:
              name: "Production API Key"
              permissions:
                - "escrow:create"
                - "escrow:read"
      responses:
        '201':
          description: API key generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/InstitutionApiKeyFull'
                  message:
                    type: string
                    example: "Store the key securely. It will not be shown again."
                  timestamp:
                    type: string
                    format: date-time
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags:
        - Institution Settings
      summary: List API keys
      description: List all API keys for the authenticated client (keys are masked).
      security:
        - BearerAuth: []
      responses:
        '200':
          description: API keys listed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/InstitutionApiKeySummary'
                  timestamp:
                    type: string
                    format: date-time
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/institution/api-keys/{id}:
    delete:
      tags:
        - Institution Settings
      summary: Revoke API key
      description: Permanently revoke an API key. This action cannot be undone.
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: API key ID to revoke
      responses:
        '200':
          description: API key revoked
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "API key revoked"
                  timestamp:
                    type: string
                    format: date-time
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ── Institution Files ──────────────────────────────────────────────────
...
  /api/v1/institution/settings/wallets/manage:
    put:
      tags:
        - Institution Settings
      summary: Add or update a wallet
      description: |
        Add a new wallet or update an existing one. If `id` is provided, updates that wallet.
        Setting `isPrimary` or `isSettlement` to true will unset the flag on other wallets.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InstitutionWalletInput'
            example:
              name: "Primary Treasury"
              address: "498GViCLvzbGnRoByJCAj7skXkAe3NBpCY2Wghcd2e4R"
              chain: "solana"
              description: "Main treasury wallet"
              isPrimary: true
      responses:
        '200':
          description: Wallet added/updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/InstitutionWalletRecord'
                  timestamp:
                    type: string
                    format: date-time
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Not authenticated
  /api/v1/institution/settings/wallets/list:
    get:
      tags:
        - Institution Settings
      summary: List wallets
      description: List all wallets for the authenticated client, ordered by primary status.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Wallet list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/InstitutionWalletRecord'
                  timestamp:
                    type: string
                    format: date-time
        '401':
          description: Not authenticated
  /api/v1/institution/settings/wallets/{id}:
    delete:
      tags:
        - Institution Settings
      summary: Delete a wallet
      description: Remove a wallet from the client's account.
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: Wallet ID
      responses:
        '200':
          description: Wallet deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Not authenticated

  # ── AI Analysis (renamed routes) ────────────────────────────────────
...
components:
...
  schemas:
...
    InstitutionWalletInput:
      type: object
      required:
        - name
        - address
      properties:
        id:
          type: string
          format: uuid
          description: Wallet ID (if updating an existing wallet)
        name:
          type: string
          description: Human-readable wallet name
          example: "Primary Treasury"
        address:
          type: string
          description: Solana wallet address
          example: "498GViCLvzbGnRoByJCAj7skXkAe3NBpCY2Wghcd2e4R"
        chain:
          type: string
          default: solana
          description: Blockchain network
        description:
          type: string
          nullable: true
          description: Optional description
        provider:
          type: string
          nullable: true
          description: Custody provider name
        isPrimary:
          type: boolean
          default: false
          description: Set as primary wallet (unsets other primaries)
        isSettlement:
          type: boolean
          default: false
          description: Set as settlement wallet (unsets other settlements)
    InstitutionWalletRecord:
      type: object
      properties:
        id:
          type: string
          format: uuid
        clientId:
          type: string
          format: uuid
        name:
          type: string
        address:
          type: string
        chain:
          type: string
        description:
          type: string
          nullable: true
        provider:
          type: string
          nullable: true
        isPrimary:
          type: boolean
        isSettlement:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # ── Institution Settings Schemas ────────────────────────────────────
    InstitutionSettings:
      type: object
      properties:
        clientId:
          type: string
          format: uuid
        companyName:
          type: string
        webhookUrl:
          type: string
          nullable: true
        wallets:
          type: object
          properties:
            payerWallet:
              type: string
              nullable: true
            recipientWallet:
              type: string
              nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    InstitutionSettingsUpdate:
      type: object
      properties:
        companyName:
          type: string
          description: Updated company name
        webhookUrl:
          type: string
          description: Webhook URL for escrow status notifications
    InstitutionWalletsUpdate:
      type: object
      properties:
        payerWallet:
          type: string
          description: Solana wallet address for sending payments
          example: "498GViCLvzbGnRoByJCAj7skXkAe3NBpCY2Wghcd2e4R"
        recipientWallet:
          type: string
          description: Solana wallet address for receiving payments
          example: "8LL197pziojWHtS3zeyJonrh1swKvMZpumfesVmDgUcZ"
    InstitutionApiKeyCreate:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Human-readable name for the API key
          example: "Production API Key"
        permissions:
          type: array
          items:
            type: string
          description: List of permissions granted to this key
          example: ["escrow:create", "escrow:read"]
    InstitutionApiKeyFull:
      type: object
      description: API key object returned on creation (includes the full key)
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        key:
          type: string
          description: The full API key (only shown once at creation time)
        permissions:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
    InstitutionApiKeySummary:
      type: object
      description: API key summary (key is masked)
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        keyPrefix:
          type: string
          description: First characters of the key for identification
          example: "iek_a1b2..."
        permissions:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        lastUsedAt:
          type: string
          format: date-time
          nullable: true

    # ── Institution Files Schemas ───────────────────────────────────────
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Validation Error"
        message:
          type: string
          example: "Invalid wallet address format"
        timestamp:
          type: string
          format: date-time
...
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for admin authentication
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from /api/v1/institution/auth/login
    SettlementAuthority:
      type: apiKey
      in: header
      name: X-Settlement-Authority-Key
      description: Settlement authority API key for release operations